+++ kibot/log.py
@@ -33,6 +33,8 @@ root_logger = None
 visual_level = None
 debug_level = 0
 stop_on_warnings = False
+dont_stop = False
+errors_ignored = False
 record_error_msgs = False
 recorded_error_msgs = []
 re_dk_client_id = re.compile("DIGIKEY_CLIENT_ID=([^']+)")
@@ -141,7 +143,10 @@ class MyLogger(logging.Logger):
     def check_warn_stop(self):
         if stop_on_warnings:
             self.error('Warnings treated as errors')
-            sys.exit(WARN_AS_ERROR)
+            if dont_stop:
+                pass
+            else:
+                sys.exit(WARN_AS_ERROR)
 
     def log(self, level, msg, *args, **kwargs):
         if level < self.getEffectiveLevel():