+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.8.5] - UNRELEASED
 ### Added
+- CLI:
+  - -F, --fail-on-ignored: to return error even when using -D (#828)
 - Schematic: Support for private fields (#812)
 - Internal templates:
   - PCB2Blender_2_17: generates a pcb2blender 2.17 file to import on Blender